[1101] refactor have_courses matcher
DRY'd custom matchers for course by moving them to a matcher folder.
Refactored the request pagination tests allowing one matcher for
request and controller spec files.
