+++ spec/spec_helper.rb
@@ -2,6 +2,7 @@ require 'support/serializer_spec_helper'
 require 'rspec-json_matchers'
 require 'pundit/rspec'
 require 'jsonapi/rspec'
+require 'support/matchers/have_courses'
 
 # This file was generated by the `rails generate rspec:install` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
